requirements.txt
pymupdf
pdf2image
pytesseract
sentence-transformers
torch
torchvision
git+https://github.com/openai/CLIP.git
pillow
chromadb
transformers
accelerate
streamlit
nltk
tqdm
umap-learn
rouge
nltk
seaborn
scikit-learn
pandas
matplotlib